pyproject.toml
[tool.poetry]
name = "discord-module"
version = "0.1.0"
description = ""
authors = ["tldrd0117 <tldrd0117@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.7.3"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
